fix: Move connections into its own component
